full Blog App

+++ post/views.py
@@ -1,29 +1,63 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
-from .forms import PostFrom
+from .forms import PostFrom, CommentForm
 from django.contrib import messages
+from django.utils.text import slugify
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.db.models import Q
 
 # Create your views here.
 
 def post_index(request):
-    posts= Post.objects.all()
+    post_lists= Post.objects.all()
+    query = request.GET.get('q')
+    if query:
+        post_lists = post_lists.filter(
+        Q(title__icontains = query)|
+        Q(content__icontains=query)|
+        Q(user__first_name__icontains=query)|
+        Q(user__last_name__icontains= query)
+        ).distinct()
+
+
+    paginator = Paginator(post_lists, 2) # Show 25 contacts per page
+
+    page = request.GET.get('page')
+    try:
+        posts = paginator.page(page)
+    except PageNotAnInteger:
+        # If page is not an integer, deliver first page.
+        posts = paginator.page(1)
+    except EmptyPage:
+        # If page is out of range (e.g. 9999), deliver last page of results.
+        posts = paginator.page(paginator.num_pages)
     return render(request, 'post/index.html', {'posts': posts})
 
-def post_detail(request, id):
-    post= get_object_or_404(Post, id=id)
+def post_detail(request, postTitleSlug):
+    
+    post= get_object_or_404(Post, postTitleSlug=postTitleSlug)
+    form = CommentForm(request.POST or None)
+    if form.is_valid():
+            comment = form.save(commit=False)
+            comment.post = post
+            comment.save()
+            return HttpResponseRedirect(post.get_absolute_url())
     context = {
         'post': post,
+        'form': form,
     }
     return render(request, 'post/detail.html', context)
 
 def post_create(request):
-    
 
-   
+    if not request.user.is_authenticated:
+        raise Http404() 
 
-    form = PostFrom(request.POST or None)
+    form = PostFrom(request.POST or None, request.FILES or None)
     if form.is_valid():
-            post = form.save()
+            post = form.save(commit=False)
+            post.user = request.user
+            post.save()
             messages.success(request, 'Başarılı Bir Şekilde Oluşturdunuz!')
             return HttpResponseRedirect(post.get_absolute_url())
     
@@ -35,9 +69,11 @@ def post_create(request):
 
     return render(request, 'post/form.html', context)
 
-def post_update(request, id):
-    post= get_object_or_404(Post, id=id)
-    form = PostFrom(request.POST or None, instance=post)
+def post_update(request, postTitleSlug):
+    if not request.user.is_authenticated:
+        raise Http404() 
+    post= get_object_or_404(Post, postTitleSlug=postTitleSlug)
+    form = PostFrom(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
             form.save()
             messages.success(request, 'Başarılı Bir Şekilde Güncellediniz!', extra_tags='mesaj-basarili')
@@ -47,8 +83,10 @@ def post_update(request, id):
     return render(request, 'post/form.html', context)
 
 
-def post_delete(request, id):
-    post= get_object_or_404(Post, id=id)
+def post_delete(request, postTitleSlug):
+    if not request.user.is_authenticated:
+        raise Http404() 
+    post= get_object_or_404(Post, postTitleSlug=postTitleSlug)
     post.delete()
     messages.success(request, 'Başarılı Bir Şekilde Sildindiniz!')
     return redirect('post:index')
